Refactoring
Combined firebaseDB and Firebase
